+++ src/docs/constants/e.js
@@ -1,0 +1,15 @@
+math.docs.e = math.docs.E = {
+    "name": "e",
+    "category": "Constants",
+    "syntax": [
+        "e"
+    ],
+    "description": "Euler's number, the base of the natural logarithm. Approximately equal to 2.71828",
+    "examples": [
+        "e",
+        "e ^ 2",
+        "exp(2)",
+        "log(e)"
+    ],
+    "seealso": ["exp"]
+};