+++ src/docs/function/arithmetic/multiply.js
@@ -1,0 +1,19 @@
+math.docs.multiply = {
+    "name": "multiply",
+    "category": "Operators",
+    "syntax": [
+        "x * y",
+        "multiply(x, y)"
+    ],
+    "description": "multiply two values.",
+    "examples": [
+        "2.1 * 3.6",
+        "ans / 3.6",
+        "2 * 3 + 4",
+        "2 * (3 + 4)",
+        "3 * 2.1 km"
+    ],
+    "seealso": [
+        "divide"
+    ]
+};