+++ src/docs/function/arithmetic/unaryminus.js
@@ -1,0 +1,17 @@
+math.docs.unaryminus = {
+    "name": "unaryminus",
+    "category": "Operators",
+    "syntax": [
+        "-x",
+        "unaryminus(x)"
+    ],
+    "description":
+        "Inverse the sign of a value.",
+    "examples": [
+        "-4.5",
+        "-(-5.6)"
+    ],
+    "seealso": [
+        "add", "subtract"
+    ]
+};