+++ src/docs/function/matrix/det.js
@@ -1,0 +1,15 @@
+math.docs.det = {
+    "name": "det",
+    "category": "Numerics",
+    "syntax": [
+        "det(x)"
+    ],
+    "description": "Calculate the determinant of a matrix",
+    "examples": [
+        "det([1, 2; 3, 4])",
+        "det([-2, 2, 3; -1, 1, 3; 2, 0, -1])"
+    ],
+    "seealso": [
+        "concat", "diag", "eye", "inv", "range", "size", "squeeze", "transpose", "zeros"
+    ]
+};