+++ src/docs/function/matrix/transpose.js
@@ -1,0 +1,17 @@
+math.docs.transpose = {
+    "name": "transpose",
+    "category": "Numerics",
+    "syntax": [
+        "x'",
+        "transpose(x)"
+    ],
+    "description": "Transpose a matrix",
+    "examples": [
+        "a = [1, 2, 3; 4, 5, 6]",
+        "transpose(a)",
+        "a'"
+    ],
+    "seealso": [
+        "concat", "det", "diag", "eye", "inv", "ones", "range", "size", "squeeze", "zeros"
+    ]
+};