+++ src/docs/function/probability/random.js
@@ -1,0 +1,14 @@
+math.docs.random = {
+    "name": "random",
+    "category": "Probability",
+    "syntax": [
+        "random()"
+    ],
+    "description":
+        "Return a random number between 0 and 1.",
+    "examples": [
+        "random()",
+        "100 * random()"
+    ],
+    "seealso": []
+};