+++ src/docs/function/utils/format.js
@@ -1,0 +1,14 @@
+math.docs.format = {
+    "name": "format",
+    "category": "Utils",
+    "syntax": [
+        "format(value)"
+    ],
+    "description": "Format a value of any type as string.",
+    "examples": [
+        "format(2.3)",
+        "format(3 - 4i)",
+        "format([])"
+    ],
+    "seealso": []
+};