+++ src/docs/function/utils/help.js
@@ -1,0 +1,13 @@
+math.docs.help = {
+    "name": "help",
+    "category": "Utils",
+    "syntax": [
+        "help(object)"
+    ],
+    "description": "Display documentation on a function or data type.",
+    "examples": [
+        "help(\"sqrt\");",
+        "help(\"complex\");"
+    ],
+    "seealso": []
+};