+++ src/docs/function/utils/import.js
@@ -1,0 +1,13 @@
+math.docs['import'] = {
+    "name": "import",
+    "category": "Utils",
+    "syntax": [
+        "import(string)"
+    ],
+    "description": "Import functions from a file.",
+    "examples": [
+        "import(\"numbers\")",
+        "import(\"./mylib.js\")"
+    ],
+    "seealso": []
+};